Importiert Daten

Diese Funktionen importiert Daten gemäss einer Konfiguration in einen lokalen rolf rds store. Damit der Import durchgeführt werden kann, müssen folgende Anforderungen erfüllt sein:

Usage

import(config)

Argumente

config

Liste mit sämtlichen Konfigurationsoptionen

Rückgabewert

invisible(NULL)

Details

Eine Lesefunktion für die Daten muss entweder als Funktion in einem Package oder im globalen Enviroment existieren

Der Store enthält die notwendigen Metadaten

Aktuell werden als Quelle curl Downloads (ftp, http) und das Dateisystem unterstütz.

config

Die Konfiguration wird am einfachsten in einer externen yaml Datei gespeichert. Eine Beschreibung der verschiedenen Optionen ist in der Beispieldatei config_mythenquai.yml enthalten.

logging

Das Packaging nutzt einen lgr::lgr Package Level Logger mit dem Namen "rOstluft.import". Mehr Infos in der lgr Vignette

Beispiele

lg <- lgr::get_logger("rOstluft.import")
lg$set_threshold("trace")

config_file <- system.file("extdata", "config_mythenquai.yml", package = "rOstluft.import")
config <- yaml::read_yaml(config_file)
import(config)
#> INFO  [23:14:35.026] Starte import 
#> INFO  [23:14:35.073] Kein Download notwendig. Alle Dateien bereits gecached oder importiert. 
#> INFO  [23:14:35.085] Lese Datei: tmp/cache/seepolizei/mythenquai_2022-03-16.json 
#> DEBUG [23:14:35.296] Wende Metainformationen an 
#> TRACE [23:14:35.300] data$site = meta[meta$site_short == data$site]$site 
#> TRACE [23:14:35.307] data$unit = meta[meta$parameter_original == data$parameter]$unit 
#> TRACE [23:14:35.313] data$parameter = meta[meta$parameter_original == data$parameter]$parameter 
#> INFO  [23:14:35.434] Schreibe importierte Daten in Datei: tmp/imported/seepolizei/2022/mythenquai_2022-03-16.rds 
#> INFO  [23:14:35.440] Lösche Cache Datei: tmp/cache/seepolizei/mythenquai_2022-03-16.json 
#> INFO  [23:14:35.443] Lese Datei: tmp/cache/seepolizei/mythenquai_2022-03-17.json 
#> DEBUG [23:14:35.504] Wende Metainformationen an 
#> TRACE [23:14:35.506] data$site = meta[meta$site_short == data$site]$site 
#> TRACE [23:14:35.513] data$unit = meta[meta$parameter_original == data$parameter]$unit 
#> TRACE [23:14:35.520] data$parameter = meta[meta$parameter_original == data$parameter]$parameter 
#> INFO  [23:14:35.597] Schreibe importierte Daten in Datei: tmp/imported/seepolizei/2022/mythenquai_2022-03-17.rds 
#> INFO  [23:14:35.603] Lösche Cache Datei: tmp/cache/seepolizei/mythenquai_2022-03-17.json 
#> INFO  [23:14:35.607] Lese Datei: tmp/cache/seepolizei/mythenquai_2022-03-18.json 
#> DEBUG [23:14:35.670] Wende Metainformationen an 
#> TRACE [23:14:35.673] data$site = meta[meta$site_short == data$site]$site 
#> TRACE [23:14:35.679] data$unit = meta[meta$parameter_original == data$parameter]$unit 
#> TRACE [23:14:35.686] data$parameter = meta[meta$parameter_original == data$parameter]$parameter 
#> INFO  [23:14:35.764] Schreibe importierte Daten in Datei: tmp/imported/seepolizei/2022/mythenquai_2022-03-18.rds 
#> INFO  [23:14:35.770] Lösche Cache Datei: tmp/cache/seepolizei/mythenquai_2022-03-18.json 
#> INFO  [23:14:35.774] Lese Datei: tmp/cache/seepolizei/mythenquai_2022-03-19.json 
#> DEBUG [23:14:35.833] Wende Metainformationen an 
#> TRACE [23:14:35.835] data$site = meta[meta$site_short == data$site]$site 
#> TRACE [23:14:35.841] data$unit = meta[meta$parameter_original == data$parameter]$unit 
#> TRACE [23:14:35.847] data$parameter = meta[meta$parameter_original == data$parameter]$parameter 
#> INFO  [23:14:35.925] Schreibe importierte Daten in Datei: tmp/imported/seepolizei/2022/mythenquai_2022-03-19.rds 
#> INFO  [23:14:35.933] Lösche Cache Datei: tmp/cache/seepolizei/mythenquai_2022-03-19.json 
#> INFO  [23:14:35.937] Import beendet